fixed bugs, added sync mechanism to wine glasses

+++ watch/app/src/main/java/com/imsproject/watch/utils/UtilFuntions.kt
@@ -45,7 +45,7 @@ fun cartesianToPolar(x: Double, y:Double) : Pair<Float,Float> {
         atan2(
             y - SCREEN_CENTER.y,
             x - SCREEN_CENTER.x
-        ).toDouble()
+        )
     ).toFloat()
     return distance to angle
 }
